+++ resources/views/components/notification/link-tag.blade.php
@@ -1,0 +1,8 @@
+@props([
+    'id' => null,
+    'markAsReadOnClick' => false,
+])
+
+<a {{ $attributes }} {!! (! empty($id) && $markAsReadOnClick ? 'wire:click="markAsRead(\'' . $id . '\')"' : '') !!}>
+    {{ $slot }}
+</a>